# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Share2People API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/shares": {
            "get": {
                "tags": [
                    "Shares"
                ],
                "summary": "Get all shares",
                "description": "Get all shares.",
                "operationId": "4ece1547770615e0691697d836103c3f",
                "responses": {
                    "200": {
                        "description": "List of shares",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "My Share"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Optional description"
                                            },
                                            "content_type": {
                                                "type": "string",
                                                "example": "markdown"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "## Markdown Content"
                                            },
                                            "file_url": {
                                                "type": "string",
                                                "example": "https://example.com/image.jpg"
                                            },
                                            "short_code": {
                                                "type": "string",
                                                "example": "abc123"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2025-02-20 12:00:00"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "example": "2025-02-20 12:00:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Shares"
                ],
                "summary": "Create a new share",
                "description": "Create a new share.",
                "operationId": "61c991abffa2190f4b7fa948c46db043",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "title",
                                    "content_type"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "My Share"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Optional description"
                                    },
                                    "content_type": {
                                        "type": "string",
                                        "enum": [
                                            "shortlink",
                                            "link",
                                            "code",
                                            "markdown",
                                            "image"
                                        ],
                                        "example": "markdown"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "## Markdown Content"
                                    },
                                    "file_url": {
                                        "type": "string",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "short_code": {
                                        "type": "string",
                                        "example": "abc123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Share created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "My Share"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Optional description"
                                        },
                                        "content_type": {
                                            "type": "string",
                                            "example": "markdown"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "## Markdown Content"
                                        },
                                        "file_url": {
                                            "type": "string",
                                            "example": "https://example.com/image.jpg"
                                        },
                                        "short_code": {
                                            "type": "string",
                                            "example": "abc123"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2025-02-20 12:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2025-02-20 12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/shares/{id}": {
            "get": {
                "tags": [
                    "Shares"
                ],
                "summary": "Get share by ID",
                "description": "Get a share by ID.",
                "operationId": "00cff882d9dc606efaf088c9864cc9d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Share ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Share details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "My Share"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Optional description"
                                        },
                                        "content_type": {
                                            "type": "string",
                                            "example": "markdown"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "## Markdown Content"
                                        },
                                        "file_url": {
                                            "type": "string",
                                            "example": "https://example.com/image.jpg"
                                        },
                                        "short_code": {
                                            "type": "string",
                                            "example": "abc123"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2025-02-20 12:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2025-02-20 12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Shares"
                ],
                "summary": "Update an existing share",
                "description": "Update an existing share.",
                "operationId": "686b42f95890154784eb4e4fd6be6e5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Share ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content_type"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Share"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "content_type": {
                                        "type": "string",
                                        "enum": [
                                            "shortlink",
                                            "link",
                                            "code",
                                            "markdown",
                                            "image"
                                        ],
                                        "example": "markdown"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "## Updated Content"
                                    },
                                    "file_url": {
                                        "type": "string",
                                        "example": "https://example.com/updated.jpg"
                                    },
                                    "short_code": {
                                        "type": "string",
                                        "example": "upd123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Share updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Updated Share"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Updated description"
                                        },
                                        "content_type": {
                                            "type": "string",
                                            "example": "markdown"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "## Updated Content"
                                        },
                                        "file_url": {
                                            "type": "string",
                                            "example": "https://example.com/updated.jpg"
                                        },
                                        "short_code": {
                                            "type": "string",
                                            "example": "upd123"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2025-02-20 12:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2025-02-20 12:30:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Shares"
                ],
                "summary": "Delete a share",
                "description": "Delete a share.",
                "operationId": "a5a7569fd2ad5479dc5859334d295515",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Share ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Share deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}